rm warning

+++ QiniuSDK/QiniuHttpClient.m
@@ -103,7 +103,7 @@
 - (NSDictionary *)convertToPostParams{
     NSMutableDictionary *params = [NSMutableDictionary dictionaryWithDictionary:self.params];
     if (self.checkCrc == 1) {
-        params[@"crc32"] = [NSString stringWithFormat:@"%ld", self.crc32];
+        params[@"crc32"] = [NSString stringWithFormat:@"%u", (unsigned int)self.crc32];
     }
     return params;
 }
